Add files via upload
This project consists of a model developed using machine learning, to scrape data from a specific app review from Google Play Store, and based on a set of predetermined answers, the model will automate replies to the reviews. This project aims to enhance user engagement and customer support for mobile apps by automating the response process to user reviews while ensuring that feedback is acknowledged and addressed effectively.